split meta optimizer files
